fix distilbert support
