+++ src/transformer_deploy/backends/pytorch_utils.py
@@ -15,12 +15,15 @@
 """
 Utils related to Pytorch inference.
 """
-
+from collections import OrderedDict
 from typing import Callable, Dict
+from typing import OrderedDict as Od
+from typing import Tuple
 
 import numpy as np
 import torch
-from transformers import PreTrainedModel
+from torch.onnx import TrainingMode
+from transformers import AutoConfig, PreTrainedModel
 
 
 def infer_classification_pytorch(
@@ -59,3 +62,65 @@ def infer_feature_extraction_pytorch(
         return model_output
 
     return infer
+
+
+def get_model_size(path: str) -> Tuple[int, int]:
+    """
+    Find number of attention heads and hidden layer size of a model
+    :param path: path to model
+    :return: tupple of # of attention heads and hidden layer size (0 if not found)
+    """
+    config = AutoConfig.from_pretrained(pretrained_model_name_or_path=path)
+    num_attention_heads = getattr(config, "num_attention_heads", 0)
+    hidden_size = getattr(config, "hidden_size", 0)
+    return num_attention_heads, hidden_size
+
+
+def convert_to_onnx(
+    model_pytorch: PreTrainedModel,
+    output_path: str,
+    inputs_pytorch: Od[str, torch.Tensor],
+    opset: int = 12,
+) -> None:
+    """
+    Convert a Pytorch model to an ONNX graph by tracing the provided input inside the Pytorch code.
+    :param model_pytorch: Pytorch model (transformers)
+    :param output_path: where to save ONNX file
+    :param inputs_pytorch: Tensor, can be dummy data, shape is not important as we declare all axes as dynamic.
+    Should be on the same device than the model (CPU or GPU)
+    :param opset: version of ONNX protocol to use, usually 12, or 13 if you use per channel quantized model
+    """
+    # dynamic axis == variable length axis
+    dynamic_axis = OrderedDict()
+    for k in inputs_pytorch.keys():
+        dynamic_axis[k] = {0: "batch_size", 1: "sequence"}
+    dynamic_axis["output"] = {0: "batch_size"}
+    with torch.no_grad():
+        torch.onnx.export(
+            model_pytorch,  # model to optimize
+            args=tuple(inputs_pytorch.values()),  # tuple of multiple inputs
+            f=output_path,  # output path / file object
+            opset_version=opset,  # the ONNX version to use, 13 if quantized model, 12 for not quantized ones
+            do_constant_folding=True,  # simplify model (replace constant expressions)
+            input_names=list(inputs_pytorch.keys()),  # input names
+            output_names=["output"],  # output axis name
+            dynamic_axes=dynamic_axis,  # declare dynamix axis for each input / output
+            training=TrainingMode.EVAL,  # always put the model in evaluation mode
+            verbose=False,
+        )
+
+
+def convert_to_quant_onnx(
+    model_pytorch: PreTrainedModel, output_path: str, inputs_pytorch: Od[str, torch.Tensor]
+) -> None:
+    """
+    Convert a quantized Pytorch model to ONNX file.
+    :param model_pytorch: Pytorch model
+    :param output_path: ONNX file path
+    :param inputs_pytorch: some dummy input (Pytorch tensor on the same device than the model)
+    """
+    from pytorch_quantization.nn import TensorQuantizer
+
+    TensorQuantizer.use_fb_fake_quant = True
+    convert_to_onnx(model_pytorch=model_pytorch, output_path=output_path, inputs_pytorch=inputs_pytorch, opset=13)
+    TensorQuantizer.use_fb_fake_quant = False
